requirements.txt
pyrofork
TgCrypto-pyrofork
python-dotenv
apscheduler